// management/frontend/src/api/schemas/menuCreate.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCreate {
  parent_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  name: string;
  type?: number;
  path?: string | null;
  icon?: string | null;
  permission_code?: string | null;
  sort?: number;
  is_active?: boolean;
  is_visible?: boolean;
  is_cache?: boolean;
}