// management/frontend/src/api/zod/system-menus/system-menus.ts
export const systemMenusCreateMenuBodyNameMax = 64;

export const systemMenusCreateMenuBodyTypeDefault = 0;
export const systemMenusCreateMenuBodyPathOneMax = 128;

export const systemMenusCreateMenuBodyIconOneMax = 64;

export const systemMenusCreateMenuBodyPermissionCodeOneMax = 128;

export const systemMenusCreateMenuBodySortDefault = 1;
export const systemMenusCreateMenuBodyIsActiveDefault = true;
export const systemMenusCreateMenuBodyIsVisibleDefault = true;
export const systemMenusCreateMenuBodyIsCacheDefault = true;

export const SystemMenusCreateMenuBody = zod.object({
  "parent_id": zod.union([zod.uuid(),zod.null()]).optional(),
  "name": zod.string().min(1).max(systemMenusCreateMenuBodyNameMax),
  "type": zod.number().default(systemMenusCreateMenuBodyTypeDefault),
  "path": zod.union([zod.string().max(systemMenusCreateMenuBodyPathOneMax),zod.null()]).optional(),
  "icon": zod.union([zod.string().max(systemMenusCreateMenuBodyIconOneMax),zod.null()]).optional(),
  "permission_code": zod.union([zod.string().max(systemMenusCreateMenuBodyPermissionCodeOneMax),zod.null()]).optional(),
  "sort": zod.number().default(systemMenusCreateMenuBodySortDefault),
  "is_active": zod.boolean().default(systemMenusCreateMenuBodyIsActiveDefault),
  "is_visible": zod.boolean().default(systemMenusCreateMenuBodyIsVisibleDefault),
  "is_cache": zod.boolean().default(systemMenusCreateMenuBodyIsCacheDefault)
})
// ...
export const systemMenusUpdateMenuBodyNameMax = 64;

export const systemMenusUpdateMenuBodyTypeDefault = 0;
export const systemMenusUpdateMenuBodyPathOneMax = 128;

export const systemMenusUpdateMenuBodyIconOneMax = 64;

export const systemMenusUpdateMenuBodyPermissionCodeOneMax = 128;

export const systemMenusUpdateMenuBodySortDefault = 1;
export const systemMenusUpdateMenuBodyIsActiveDefault = true;
export const systemMenusUpdateMenuBodyIsVisibleDefault = true;
export const systemMenusUpdateMenuBodyIsCacheDefault = true;

export const SystemMenusUpdateMenuBody = zod.object({
  "parent_id": zod.union([zod.uuid(),zod.null()]).optional(),
  "name": zod.string().min(1).max(systemMenusUpdateMenuBodyNameMax),
  "type": zod.number().default(systemMenusUpdateMenuBodyTypeDefault),
  "path": zod.union([zod.string().max(systemMenusUpdateMenuBodyPathOneMax),zod.null()]).optional(),
  "icon": zod.union([zod.string().max(systemMenusUpdateMenuBodyIconOneMax),zod.null()]).optional(),
  "permission_code": zod.union([zod.string().max(systemMenusUpdateMenuBodyPermissionCodeOneMax),zod.null()]).optional(),
  "sort": zod.number().default(systemMenusUpdateMenuBodySortDefault),
  "is_active": zod.boolean().default(systemMenusUpdateMenuBodyIsActiveDefault),
  "is_visible": zod.boolean().default(systemMenusUpdateMenuBodyIsVisibleDefault),
  "is_cache": zod.boolean().default(systemMenusUpdateMenuBodyIsCacheDefault)
})